Query BC Fish Observation Collection

Query a collection from the BC Fish Observation API.

fop_query_collection_passage(
  collection_id,
  filter = NULL,
  limit = 10000,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  sortby = NULL,
  groupby = NULL
)

Arguments

collection_id

A character string of the collection id.

filter

A named vector or list of the filter(s) to apply, where the list names correspond to column names and the list values correspond to the desired value, e.g. list(gnis_name = "Sangan River").

limit

A positive whole number that limits the number of features in the response.

offset

A positive whole number that starts the response at an offset.

bbox

A vector of four numbers indicating bounding box in lon/lat (e.g. c(minLon, minLat, maxLon, maxLat)) that features in response must intersect.

properties

A vector of strings of the column names to include. If NULL (default), all columns are retained.

transform

A character vector with the name of the valid transform function followed by the parameter values (e.g. c("ST_Simplify", 100)).

sortby

A string that sorts the response items by a property. Default is ascending but property name can be prepended with '-' to indicate descending.

groupby

A string of the property name (i.e., column name) to group response features by. This would typically be used with an aggregate transform function.

See also

Examples

library(sf)
collection_id <- "bcfishpass.barriers_falls"
filter <- list(barrier_type = "FALLS")
fop_query_collection_passage(collection_id, filter = filter)
#> Simple feature collection with 159 features and 142 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -127.6438 ymin: 49.01794 xmax: -114.1039 ymax: 55.23527
#> Geodetic CRS:  WGS 84
#> # A data frame: 159 × 143
#>    accessibility_model_salmon accessibility_m… accessibility_m… all_rearing_bel…
#>    <chr>                      <chr>            <chr>                       <dbl>
#>  1 NA                         NA               NA                              0
#>  2 NA                         NA               NA                              0
#>  3 NA                         NA               NA                              0
#>  4 NA                         NA               NA                              0
#>  5 NA                         NA               NA                              0
#>  6 NA                         NA               NA                              0
#>  7 NA                         NA               NA                              0
#>  8 ACCESSIBLE                 NA               NA                              0
#>  9 NA                         NA               NA                              0
#> 10 NA                         NA               NA                              0
#> # … with 149 more rows, and 139 more variables: all_rearing_km <dbl>,
#> #   all_spawning_belowupstrbarriers_km <dbl>, all_spawning_km <dbl>,
#> #   all_spawningrearing_belowupstrbarriers_km <dbl>,
#> #   all_spawningrearing_km <dbl>, barrier_name <chr>, barrier_type <chr>,
#> #   barriers_falls_id <dbl>, blue_line_key <dbl>,
#> #   ch_rearing_belowupstrbarriers_km <dbl>, ch_rearing_km <dbl>,
#> #   ch_spawning_belowupstrbarriers_km <dbl>, ch_spawning_km <dbl>, …