Query BC Fish Observation Collection

Query a collection from the BC Fish Observation API.

fop_query_collection_observation(
  collection_id,
  filter = NULL,
  limit = 10000,
  offset = 0,
  bbox = NULL,
  properties = NULL,
  transform = NULL,
  sortby = NULL,
  groupby = NULL
)

Arguments

collection_id

A character string of the collection id.

filter

A named vector or list of the filter(s) to apply, where the list names correspond to column names and the list values correspond to the desired value, e.g. list(gnis_name = "Sangan River").

limit

A positive whole number that limits the number of features in the response.

offset

A positive whole number that starts the response at an offset.

bbox

A vector of four numbers indicating bounding box in lon/lat (e.g. c(minLon, minLat, maxLon, maxLat)) that features in response must intersect.

properties

A vector of strings of the column names to include. If NULL (default), all columns are retained.

transform

A character vector with the name of the valid transform function followed by the parameter values (e.g. c("ST_Simplify", 100)).

sortby

A string that sorts the response items by a property. Default is ascending but property name can be prepended with '-' to indicate descending.

groupby

A string of the property name (i.e., column name) to group response features by. This would typically be used with an aggregate transform function.

See also

Examples

library(sf)
collection_id <- "bcfishobs.fiss_fish_obsrvtn_events_sp"
filter <- list(species_code = "CO", watershed_group_code = "CLAY")
fop_query_collection_observation(collection_id, filter = filter)
#> Simple feature collection with 387 features and 17 fields
#> Geometry type: POINT
#> Dimension:     XYZ
#> Bounding box:  xmin: -126.4678 ymin: 49.01367 xmax: -125.3825 ymax: 49.64545
#> z_range:       zmin: 1 zmax: 553.003
#> Geodetic CRS:  WGS 84
#> # A data frame: 387 × 18
#>    agency_id agency_name       blue_line_key distance_to_str… downstream_route_…
#>    <chr>     <chr>                     <dbl>            <dbl>              <dbl>
#>  1 NA        NA                    354088196           57.7                 78.3
#>  2 NA        NA                    354088366           12.0                  0  
#>  3 NA        NA                    354088440          163.                   0  
#>  4 NA        Consultants - Ge…     354099135          250.                   0  
#>  5 NA        NA                    354099837           46.7                293. 
#>  6 NA        NA                    354099837           41.6                362. 
#>  7 NA        NA                    354104075            3.27              1248. 
#>  8 6180      Central Westcoas…     354104110            0.354             1189. 
#>  9 6180      Central Westcoas…     354104110            0.186             1227. 
#> 10 NA        NA                    354107667           83.4                501. 
#> # … with 377 more rows, and 13 more variables: fish_observation_point_id <dbl>,
#> #   linear_feature_id <dbl>, localcode_ltree <chr>, match_type <chr>,
#> #   observation_date <chr>, source <chr>, source_ref <chr>, species_code <chr>,
#> #   species_id <dbl>, waterbody_key <dbl>, watershed_group_code <chr>,
#> #   wscode_ltree <chr>, geometry <POINT [°]>